Reorganize command modules
