+++ cmd/mailroom/main.go
@@ -1,21 +1,10 @@
 package main
 
 import (
-	ulog "log"
 	"log/slog"
 	"os"
-	"os/signal"
-	goruntime "runtime"
-	"syscall"
-	"time"
 
-	"github.com/getsentry/sentry-go"
-	_ "github.com/lib/pq"
-	"github.com/nyaruka/gocommon/uuids"
-	"github.com/nyaruka/mailroom"
-	"github.com/nyaruka/mailroom/runtime"
-	slogmulti "github.com/samber/slog-multi"
-	slogsentry "github.com/samber/slog-sentry/v2"
+	"github.com/nyaruka/mailroom/cmd"
 
 	_ "github.com/nyaruka/mailroom/core/runner/handlers"
 	_ "github.com/nyaruka/mailroom/core/runner/hooks"
@@ -50,79 +39,8 @@ var (
 )
 
 func main() {
-	cfg, err := runtime.LoadConfig()
-	if err != nil {
-		slog.Error("error creating runtime", "error", err)
+	if err := cmd.Service(version, date); err != nil {
+		slog.Error(err.Error())
 		os.Exit(1)
-	}
-	cfg.Version = version
-
-	// configure our logger
-	logHandler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: cfg.LogLevel})
-	slog.SetDefault(slog.New(logHandler))
-
-	// if we have a DSN entry, try to initialize it
-	if cfg.SentryDSN != "" {
-		err := sentry.Init(sentry.ClientOptions{Dsn: cfg.SentryDSN, ServerName: cfg.InstanceID, Release: version, AttachStacktrace: true})
-		if err != nil {
-			slog.Error("error initiating sentry client", "error", err, "dsn", cfg.SentryDSN)
-			os.Exit(1)
-		}
-
-		defer sentry.Flush(2 * time.Second)
-
-		slog.SetDefault(slog.New(
-			slogmulti.Fanout(
-				logHandler,
-				slogsentry.Option{Level: slog.LevelError}.NewSentryHandler(),
-			),
-		))
-	}
-
-	log := slog.With("comp", "main")
-	log.Info("starting mailroom", "version", version, "released", date)
-
-	if cfg.UUIDSeed != 0 {
-		uuids.SetGenerator(uuids.NewSeededGenerator(int64(cfg.UUIDSeed), time.Now))
-		log.Warn("using seeded UUID generation", "uuid-seed", cfg.UUIDSeed)
-	}
-
-	rt, err := runtime.NewRuntime(cfg)
-	if err != nil {
-		slog.Error("error creating runtime", "error", err)
-		os.Exit(1)
-	}
-
-	svc := mailroom.NewService(rt)
-
-	if err := svc.Start(); err != nil {
-		log.Error("unable to start server", "error", err)
-		os.Exit(1)
-	}
-
-	// handle our signals
-	handleSignals(svc)
-}
-
-// handleSignals takes care of trapping quit, interrupt or terminate signals and doing the right thing
-func handleSignals(svc *mailroom.Service) {
-	sigs := make(chan os.Signal, 1)
-	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
-
-	for {
-		sig := <-sigs
-		log := slog.With("comp", "main", "signal", sig)
-
-		switch sig {
-		case syscall.SIGQUIT:
-			buf := make([]byte, 1<<20)
-			stacklen := goruntime.Stack(buf, true)
-			log.Info("received quit signal, dumping stack")
-			ulog.Printf("\n%s", buf[:stacklen])
-		case syscall.SIGINT, syscall.SIGTERM:
-			log.Info("received exit signal, exiting")
-			svc.Stop()
-			return
-		}
 	}
 }
